CORE-13661 Separate demo tests
