+++ e2e-tests/build.gradle
@@ -67,6 +67,12 @@ dependencies {
     cpis project(path: ':e2e-tests:cpbs:ledger-utxo-advanced-chainable-test-app', configuration: 'cordaCPB')
     cpis project(path: ':e2e-tests:cpbs:ledger-utxo-advanced-ownable-test-app', configuration: 'cordaCPB')
 
+    cpis project(path: ':examples:ledger-utxo-advanced-fungible-demo-app', configuration: 'cordaCPB')
+    cpis project(path: ':examples:ledger-utxo-advanced-identifiable-demo-app', configuration: 'cordaCPB')
+    cpis project(path: ':examples:ledger-utxo-advanced-issuable-demo-app', configuration: 'cordaCPB')
+    cpis project(path: ':examples:ledger-utxo-advanced-chainable-demo-app', configuration: 'cordaCPB')
+    cpis project(path: ':examples:ledger-utxo-advanced-ownable-demo-app', configuration: 'cordaCPB')
+
     e2eTestImplementation "org.jetbrains.kotlin:kotlin-stdlib:$kotlinVersion"
     e2eTestImplementation "com.fasterxml.jackson.module:jackson-module-kotlin:$jacksonVersion"
     e2eTestImplementation "org.assertj:assertj-core:$assertjVersion"
