+++ e2e-tests/src/e2eTest/kotlin/com/r3/corda/ledger/utxo/e2etest/OwnableTests.kt
@@ -108,9 +108,6 @@ class OwnableTests {
         assertThat(response.before).isEmpty()
         assertThat(response.after).hasSize(2)
         assertThat(response.consumed).isEmpty()
-//        assertThat(response.after).hasSize(response.before.size + 2)
-//        assertThat(response.after).containsAll(response.before)
-//        assertThat(response.consumed - response.before).isEmpty()
     }
 
     @Test
